+++ src/ambition/version_details.cc
@@ -18,6 +18,12 @@ std::vector<VersionUpdates> versionDetails = {
     .identifier = IDENTIFIER_PREVIEW,
     .sections = {
       {
+        .heading = HEADING_ADDED,
+        .items = {
+          _("Added a feedback sound when clicking the Clone Caravan button."),
+        },
+      },
+      {
         .heading = HEADING_FIXED,
         .items = {
           _("Fixed issue with \"Building Complete\" (and other) sounds not being audible when they are far away from the current screen location because they became too quiet."),